Fix shortage and volatile
Shortage has its own value for `type`. In addition marginal cost is set
to a high value. In our example its the most expensive option to balance
demand and generation.

Set fixed to true. As a consequence hourly renewable generation is set
exogenous; the product of capacity multiplied by the associated
generation profile.