Fix transmission line efficiencies
They where assigned from the viewpoint of the `Bus`, e.g. the efficiency
of the energy flow of a `source` `Bus` to a transmission line used the
region's "output ratio". This is wrong, as these ratios are specified
from the viewpoint of the transmission line. This means for the example
above, that the output from a source into the transmission line is
viewed as an input to the transmission line and should thus use the
transmission line's "input ratio" for the source region.
Basically this means that the ratios had to be swapped, which this
commit does.