Fix transmission line capacity assignments
Having an "installed capacity" on both, the ingoing as well as the
outgoing energy flow doesn't make sense if it's the same value. The
value should be tied to one of them. I chose having the installed
capacity on the input ratio for now, but I need to double check whether
that is correct.