Fix transmission line capacity usage
Transmission lines gathered from the data are modelled as two separate
`Transformer`s, one for each direction in which electricity can be
transmitted. This lead to a bug, where the full installed capacity could
be used in both directions in the same time step.
In order to fix that, an additional `Sink` has to be introduced,
collecting the outputs of both `Transformer`s. Afterwards, it's possible
to simply limit the electricity collected by this `Sink` by putting the
`"installed capacity"` limit on the `Flow` into this `Sink`.
One beneficial side effect of this commit is that the "name" attribute
of the `Sink` gathering the "energy flow" is now appropriately named, so
the entry in the "parameter_name" column of the output CSV files is
specified at one point only.