Implement "DE" wide "waste" and "CO2" limits
Two things had to happen for this. First, we needed a `Flow` on which to
put the limit. This is done by adding a `Sink` after the "waste" and
"co2" buses where the `Flow` into the sink is limited according to the
data.
The second thing which had to happen, was to expand fueled `Source`s
into a triple consisting of the original `Source`, an auxiliary `Bus`
and an auxiliary `Transformer`. The auxiliary `Transformer` acts as a
splitter, applying different weights to the `Source`'s output and
splitting it into various buses. That way, the "electricity generation"
can be funnelled into the electricity bus, the CO2 emissions can be
calculated and funnelled into the CO2 bus and the waste consumption, if
any, can be calculated and funnelled into the corresponding bus.
Note how the flows are limited by using `summed_max`, but that this
parameter does not work without also specifying a `nominal_value`.