Change how PV output is collected per region
First, each region's PV bus is now created together with all the other
buses. Then, instead of only connecting each PV technology per region to
a transformer, now every source is connected to a transformer, which
then splits its output into the regions electricity bus and the PV bus,
if the corresponding source is PV. This means that there is no longer a
single PV flow to a PV bus per region. In order to still be able to
implement region wise PV expansion limits later, this means that the PV
bus now has to have a sink attached, whose flow can later be limited.