Implement renewable share ratio
It's a bit weird, but the implementation is based on the following
reasoning. If the total electricity generation is calculated by adding
renewable sources R and conventional sources C, then a renewable ratio
of r would mean:

  R >= r(R+C)

which means:

  0 >= rC + (r-1)R

So if every conventional generation is multiplied with funnelled into
the renewable share bus while being multiplied with r and every
renewable generation is put into this bus while being multiplied with
(r-1), the bus will always have a negative balance, if the renewable
share is at least met. In order to allow a bigger renewable share, the
bus can be balanced out by a positive source.