Implement investments
Using the `invest` function, a given set of parameters is checked for
whether the associated object should be created with an `investment`
parameter or not. Depending on its argument, the `invest` function
returns a dictionary which can be inserted into a `Flow` or `Storage`s
constructor call via double splats. This is why the `invest` function is
used to replace most `nominal_value=` lines with `**invest(parameters)`.

In order for this to work, the lines getting the parameters from the
mappings parsed out of the JSON file had to be modified to no longer
search for an `"installed capacity"` parameter, because this is only
present in nodes to be invested in which already have an existing
capacity.