Fix renewable share ratio implementation
Using negative output conversion factors apparently did not work. So
instead, the total demand has to be calculated in advance and be used as
the nominal value for a sink collecting renewables. The `summed_min`
parameter can then be used, to make sure that the flow into the sink
does not drop below the specified share.