Promote `Label` from `namedtuple` to `class`
This is necessary in order to get a custom string representation. At
least that was the first option I could think of. Having a custom string
representation proved necessary in order to save the result dataframe to
a file and read it back on a different machine later. For some reason
the `Label` objects could not be pickled, so they needed to be converted
to strings before saving and be parsed back to `Label`s after reading
the dataframe back in.
Promoting the `Label` to a `class` also uncovers the similarity between
`Label`s and `Key`s, so it's easier to unify both later.