Store the results in a file
At least the time dependent components of the results. They are
converted to a dataframe using `rs2df` and the return value is saved in
an HDF5 file. Since Python was happy pickling the `Label` objects, but
always raised an exception when unpickling them somewhere else, the
components of each column are converted to strings before they are
written to disk.