Generate a Sankey diagram from the results
The diagram is generated from the results dataframe created via `rs2df`.
At least from the dataframe that is read back in from the HDF5 file. I
don't know whether it also works with a raw dataframe where the columns
are tuples of objects instead of tuples of strings. It probably won't
work, but that's something that can be worked out later.
Also, the diagram can still be improved, but it suffices for me for
debugging purposes for now.