Add limit buses and sinks to the energy system
These where not part of the chain followed from the sources, as that
chain stopped at the auxiliary transformer. Without adding those
components to the energy system, the accompanying constraints did not
get build, so it's only natural, that the limits set by the flows from
the `limit_buses` to the `limit_sinks` where not respected.
This should finally fix the commodity limits not being respected by the
optimisation model.