Separate `main` from the CLI command
After decorating a function as a `@click.command`, it can't be called
that easily from an interactive REPL, because click changes the way the
function receives it's arguments. Therefore this commit separates the
decorated function from the one called internally after parsing the
command line. That way one can call `main` from a REPL the expected way.