Fix calling `reduce` on possibly empty `emissions`
When switching to only put scalar values strictly larger than zero into
the reports, the fact that this meant that the list of `emissions` could
potentially be empty was not taken into account. Since `reduce` was
called on this list without a default value, having no `emissions`
resulted in an error. The case of having no `emissions` also means that
no `emissions` have to be summed up, so `reduce` doesn't have to be
called, which is exactly the solution this commit implements. Apart from
one more indentation level due to the additional check, the code is not
changed.