Fix the relationships of a storage's capacities
Because they where put into the `optionals` dictionary, the
`invest_relation_input_capacity` and `invest_relation_output_capacity`
parameters where wrongfully added to the `Investment` object put on the
storage, instead of being applied to the storage itself.
Now these parameters are put into their own dictionary, if they are
applicable that is, and this dictionary is then expanded into the
storage construction parameters, where it belongs.