Implement configurable penalty costs
At least for transmission losses and storage inputs. The incurred costs
will be removed from the reported costs as they are only used to control
usage of transmission lines and storages. They can for example be used
to prevent transmission lines and storages from being used as a
curtailment substitute.