Add a `lifetime` for `("photovoltaics", "unknown")`
This wasn't necessary when dealing with only one year, since "lifetime"
was only used for economic calculations. Now that "lifetime" is also
used to decommission `Investment`s, the `("photovoltaics", "unknown")`
technology needs to have one, too. Since this technology is not part of
the input data but instead artificially created in order group other
"photovoltaics" technologies, the lifetime can not be determined from
the input date and is set to zero because the lifetimes of the
subordinate technologies may vary and that way they are never masked by
`("photovoltaics", "unknown")`'s lifetime.